fix(ecg): widen AccessHints.mask to uint32_t for 16/32-bit masks
The mask field was uint8_t, truncating 16/32-bit masks to 8 bits.
This caused resolvePrefetchTarget() to always return UINT32_MAX
for wider masks because the prefetch bits (shifted above bit 8)
were lost during truncation.

Changes:
- AccessHints.mask: uint8_t → uint32_t (supports all mask widths)
- SIM_CACHE_READ_MASKED: cast to uint32_t instead of uint8_t
- SIM_CACHE_READ_MASKED_PREFETCH: cast to uint32_t
- PR kernel: use computeVertexMasks() (uint32_t) for mask computation,
  keep 8-bit array for backward-compatible initMaskArray8()

This fix enables prefetch targets encoded in bits 8+ to be properly
decoded and resolved, allowing 16-bit (64 targets) and 32-bit
(direct vertex ID) prefetch modes to function.

+++ bench/include/cache_sim/graph_sim.h
@@ -85,7 +85,7 @@ private:
 // mask_val = pre-encoded mask entry from the parallel mask array.
 #define SIM_CACHE_READ_MASKED(cache, arr, idx, graph_ctx, mask_val) \
     do { \
-        (graph_ctx).hints_for_thread().mask = static_cast<uint8_t>(mask_val); \
+        (graph_ctx).hints_for_thread().mask = static_cast<uint32_t>(mask_val); \
         (cache).access(reinterpret_cast<uint64_t>(&(arr)[idx]), false); \
     } while(0)
 
@@ -96,7 +96,7 @@ private:
 // as a demand access — prefetch misses don't inflate the miss rate.
 #define SIM_CACHE_READ_MASKED_PREFETCH(cache, arr, idx, graph_ctx, mask_val) \
     do { \
-        (graph_ctx).hints_for_thread().mask = static_cast<uint8_t>(mask_val); \
+        (graph_ctx).hints_for_thread().mask = static_cast<uint32_t>(mask_val); \
         (cache).access(reinterpret_cast<uint64_t>(&(arr)[idx]), false); \
         uint32_t _pfx_target = (graph_ctx).resolvePrefetchTarget(mask_val); \
         if (_pfx_target != UINT32_MAX) { \
